+++ charts/tfserving/CHANGELOG.md
@@ -30,3 +30,4 @@
 
 * support --config-file
 * support --shell
+* support arena serve update tensorflow