+++ charts/triton/CHANGELOG.md
@@ -21,3 +21,4 @@
 
 * support --config-file
 * support --shell
+* support arena serve update triton